refactor: address code review feedback - simplify and optimize
Changes:
- Move cleanDatabase to testingFunctions.ts with IS_TEST guard
- Remove adminFunctions.ts (functions moved to appropriate modules)
- Fix N+1 query in getBooksHierarchy (use Promise.all for parallel gets)
- Remove legacy getArticles query (unused, replaced by getBooksHierarchy)
- Remove scrapeSequences wrapper function (unused legacy compatibility layer)
- Add validation to getPendingArticles (filter out entries with missing metadata)
- Use index in getBooksHierarchy for better query practice

Lines removed: ~169 lines
Performance improvements:
- N+1 query fix improves completion counting for users with many completions
- Index usage follows Convex best practices
